refactor: 에러 수정 및 파일 이동
- lint 에러 수정
- 파일 경로 수정
