+++ src/app/(MyPage)/mypage/@addReview/(.)addReview/page.tsx
@@ -1,7 +1,6 @@
 import ReviewModal from "@/components/public/modal/ReviewModal"
+import { IAddReviewPageProp } from "@/types/mypage/mypage"
 import { HydrationBoundary } from "@tanstack/react-query"
-
-import { IAddReviewPageProp } from "../../mypage"
 
 const AddReviewPage = ({ searchParams }: IAddReviewPageProp) => {
   return (
